// src/utils/enums.ts
export enum OPERATIONS {
  PLUS = '+',
  MINUS = '-',
  MULTIPLY = '×',
  DIVIDE = '÷',
}

export enum GAME_TYPE {
  MATH = 'MATH',
}

export enum GAMES {
  EQUATIONS = 'EQUATIONS',
  ARCADE = 'ARCADE',
  CALENDAR = 'CALENDAR',
  SQUARE_ROOTS = 'SQUARE ROOTS',
  SIMPLE_MATH = 'SIMPLE MATH',
}

export enum GAME_SETTINGS {
  OPERATIONS = 'OPERATIONS',
  DIFFICULTY = 'DIFFICULTY',
  DURATION = 'DURATION',
}

export enum TIMER_TYPE {
  UP = 'UP',
  DOWN = 'DOWN',
}

export enum GAME_MODE {
  PRACTICE = 'PRACTICE',
  RANKED = 'RANKED',
}

export enum ANSWER_OPTIONS {
  NUMBERS = 'NUMBERS',
  YES_NO = 'YES_NO',
  DAYS = 'DAYS',
  FOUR_OPTIONS = 'FOUR_OPTIONS',
}

// General
export enum SVGS {
  HOME = 'HOME',
  GOOGLE = 'GOOGLE',
  FACEBOOK = 'FACEBOOK',
  APPLE = 'APPLE',
  X = 'X',
  BACK = 'BACK',
}

export enum LOGIN_PROVIDERS {
  GUEST = 'GUEST',
  GOOGLE = 'GOOGLE',
  FACEBOOK = 'FACEBOOK',
  APPLE = 'APPLE',
}

export enum DAYS {
  SUNDAY = 'SUN',
  MONDAY = 'MON',
  TUESDAY = 'TUE',
  WEDNSDAY = 'WED',
  THURSDAY = 'THU',
  FRIDAY = 'FRI',
  SATURDAY = 'SAT',
}

export enum STRINGS {
  WELCOME = 'Welcome to Math Guinness',
  LOGIN_WITH = 'Log in with',
  START = 'START',
  LOGIN_MODAL_TITLE = 'Login with social',
  RANKING = 'Ranking',
  LOGOUT = 'Logout',
  PLAY_WITH_FRIENDS = 'PLAY WITH FRIENDS',
  PLAY_RANKED = 'PLAY RANKED',
}

export enum COLORS {
  WHITE = '#fff',
  MAIN_BLUE = '#2177e1',
  BLACK = '#000000',
  TEXT_MAIN = '#404040',
  TEXT_SECONDARY = '#6D7D8F',
  TEXT_LINK = '#417DFF',
  LIGHT_BLUE = '#BECFF6',
  BG_SMOKE = '#F6F8FB',
  LIGHT_GREY = '#DBDEE3',
  BG_BLUE = '#417DFF',
  RED = 'red',
  YELLOW = '#febe29',
  GREEN = '#59b2ab',
}